requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Authentication & Security
PyJWT>=2.8.0

# Environment Variables
python-dotenv>=1.0.0
python-decouple>=3.8

# Email
django-anymail>=10.2

# Filtering & Search
django-filter>=23.5

# API Documentation
drf-spectacular>=0.27.0

# CORS
django-cors-headers>=4.3.1

# Celery for Background Tasks
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
factory-boy>=3.3.0

# Code Quality
flake8>=6.1.0
black>=23.12.0

# Production Server
gunicorn>=21.2.0
